rag-backend/requirements.txt
fastapi
uvicorn
langchain
openai
PyPDF
python-dotenv
sentence-transformers
chromadb